Backend/requirements.txt
fastapi                  
uvicorn[standard]        
pillow                   
torch                    
openai                   
transformers             
python-dotenv            
python-multipart            
